Rename entry details to entryoverview
