+++ src/components/desktop/EntryPane.js
@@ -1,9 +1,9 @@
-import EntryDetails from '../entry/EntryDetails'
+import EntryOverview from '../entry/EntryOverview'
 import EntryBack from './EntryBack'
 import NavPane from './NavPane'
 
 const EntryPane = () => (
-  <NavPane nav={<EntryBack />} content={<EntryDetails />} />
+  <NavPane nav={<EntryBack />} content={<EntryOverview />} />
 )
 
 export default EntryPane
