+++ src/components/entry/EntryOverview.js
