+++ src/components/mobile/MobileLayout.js
@@ -1,7 +1,7 @@
 import { Route, Switch } from 'react-router-dom'
 
 import useRoutedTabs from '../../utils/useRoutedTabs'
-import EntryDetails from '../entry/EntryDetails'
+import EntryOverview from '../entry/EntryOverview'
 import { LocationForm } from '../form/LocationForm'
 import { PageTabs, Tab, TabList, TabPanel, TabPanels } from '../ui/PageTabs'
 import { DEFAULT_TAB, TABS } from './tabs'
@@ -41,7 +41,7 @@ const MobileLayout = () => {
             {tabPanels}
           </Route>
           <Route path="/entry/:id">
-            <EntryDetails />
+            <EntryOverview />
           </Route>
         </Switch>
       </TabPanels>
